fix: db2 correct failed master rejoin behavior
In an environment with relatively low database activity if a
master fails and rejoins before the last active log changes, the cluster
will attempt to start it as master (HADR primary). As there can only be
one master, this will cause the former primary to fail during startup.

In thenew code, a check has been added to ensure that there is no active
master in the cluster. If one is found, the former master will be started
as a slave (HADR standby).